% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_copula_gaussian_b.R
\name{cl_copula_gaussian_b}
\alias{cl_copula_gaussian_b}
\title{Clustered Copula Bayes Classifier with Gaussian marginals}
\usage{
cl_copula_gaussian_b(
  x,
  y,
  var_equal = FALSE,
  laplace = 0,
  sd_thresh = 1e-08,
  cores = 1,
  familyset = c(0, 1, 3, 4, 5, 6),
  rotations = FALSE,
  margin = 0.001,
  max_cluster = 9,
  eps = .Machine$double.eps,
  tol = c(1e-05, sqrt(.Machine$double.eps)),
  itmax = c(.Machine$integer.max, .Machine$integer.max)
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{var_equal}{asd}

\item{laplace}{asd}

\item{sd_thresh}{asd}

\item{cores}{asd}

\item{familyset}{asd}

\item{rotations}{asd}

\item{margin}{asd}

\item{max_cluster}{= 9,}

\item{eps}{= .Machine$double.eps,}

\item{tol}{= c(1.e-5, sqrt(.Machine$double.eps)),}

\item{itmax}{= c(.Machine$integer.max, .Machine$integer.max)}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Clustered Copula Bayes Classifier with Gaussian marginals.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
