% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_gaussian_nb.R, R/cl_mvgaussian_b.R,
%   R/cl_mvkernel_b.R
\name{cl_gaussian_nb}
\alias{cl_gaussian_nb}
\alias{cl_mvgaussian_b}
\alias{cl_mvkernel_b}
\title{Clustered Gaussian Naive Bayes Classifier}
\usage{
cl_gaussian_nb(
  x,
  y,
  laplace = 0,
  sd_thresh = 1e-08,
  var_equal = FALSE,
  max_cluster = 9,
  eps = .Machine$double.eps,
  tol = c(1e-05, sqrt(.Machine$double.eps)),
  itmax = c(.Machine$integer.max, .Machine$integer.max)
)

cl_mvgaussian_b(
  x,
  y,
  laplace = 0,
  Sigma_equal = FALSE,
  sd_thresh = 1e-08,
  max_cluster = 9,
  eps = .Machine$double.eps,
  tol = c(1e-05, sqrt(.Machine$double.eps)),
  itmax = c(.Machine$integer.max, .Machine$integer.max)
)

cl_mvkernel_b(
  x,
  y,
  laplace = 0,
  bw_thresh = 1e-08,
  bw_method = "nm",
  bw_adjust = 1,
  sd_thresh = 1e-08,
  max_cluster = 9,
  eps = .Machine$double.eps,
  tol = c(1e-05, sqrt(.Machine$double.eps)),
  itmax = c(.Machine$integer.max, .Machine$integer.max)
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{laplace}{asd}

\item{sd_thresh}{asd}

\item{var_equal}{asd}

\item{max_cluster}{asd}

\item{eps}{asd}

\item{tol}{asd}

\item{itmax}{asd}

\item{Sigma_equal}{asd}

\item{bw_thresh}{= 1e-8,}

\item{bw_method}{= "nm",}

\item{bw_adjust}{= 1,}
}
\value{
asd.
 \item{asd}{asd}

asd.
 \item{asd}{asd}

asd.
 \item{asd}{asd}
}
\description{
Clustered Gaussian Naive Bayes Classifier.

Clustered Gaussian Multivariate Gaussian Bayes Classifier.

Clustered Gaussian Multivariate Kernel Bayes Classifier.
}
\details{
asd

asd

asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
