% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_kernel_nb.R
\name{cl_kernel_nb}
\alias{cl_kernel_nb}
\title{Clustered Kernel Naive Bayes Classifier}
\usage{
cl_kernel_nb(
  x,
  y,
  laplace = 0,
  sd_thresh = 1e-08,
  bw_thresh = 1e-08,
  bw_method = "ucv",
  bw_adjust = 1,
  max_cluster = 9,
  eps = .Machine$double.eps,
  tol = c(1e-05, sqrt(.Machine$double.eps)),
  itmax = c(.Machine$integer.max, .Machine$integer.max)
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{laplace}{asd}

\item{sd_thresh}{asd}

\item{bw_thresh}{asd}

\item{bw_method}{asd}

\item{bw_adjust}{asd}

\item{max_cluster}{asd}

\item{eps}{asd}

\item{tol}{asd}

\item{itmax}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Clustered Kernel Naive Bayes Classifier.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
