% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_gaussian_b.R
\name{copula_gaussian_b}
\alias{copula_gaussian_b}
\title{Copula Bayes Classifier with Gaussian marginals}
\usage{
copula_gaussian_b(
  x,
  y,
  var_equal = FALSE,
  laplace = 0,
  sd_thresh = 1e-08,
  cores = 1,
  familyset = c(0, 1, 3, 4, 5, 6),
  rotations = FALSE,
  margin = 0.001
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{var_equal}{asd}

\item{laplace}{asd}

\item{sd_thresh}{asd}

\item{cores}{asd}

\item{familyset}{asd}

\item{rotations}{asd}

\item{margin}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Copula Bayes Classifier with Gaussian marginals.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
