% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_kernel_b.R
\name{copula_kernel_b}
\alias{copula_kernel_b}
\title{Copula Bayes Classifier with Kernel marginals}
\usage{
copula_kernel_b(
  x,
  y,
  sd_thresh = 1e-08,
  laplace = 0,
  bw_thresh = 1e-08,
  bw_method = "ucv",
  bw_adjust = 1,
  cores = 1,
  familyset = c(0, 1, 3, 4, 5, 6),
  rotations = FALSE,
  margin = 0.001
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{sd_thresh}{asd}

\item{laplace}{asd}

\item{bw_thresh}{asd}

\item{bw_method}{asd}

\item{bw_adjust}{asd}

\item{cores}{asd}

\item{familyset}{asd}

\item{rotations}{asd}

\item{margin}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Copula Bayes Classifier with Kernel marginals.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
