% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_by_rowsum.R
\name{divide_by_rowsum}
\alias{divide_by_rowsum}
\title{Divide row by row sum}
\usage{
divide_by_rowsum(X)
}
\arguments{
\item{X}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Divide row by row sum.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
