% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vartypes_x.R
\name{f_vartypes_x}
\alias{f_vartypes_x}
\title{Synthetic Minority Oversampling Technique (SMOTE)}
\usage{
f_vartypes_x(x)
}
\arguments{
\item{x}{asd}
}
\value{
a list with resampled dataset.
 \item{x_new}{Resampled feature matrix.}
 \item{y_new}{Resampled target variable.}
 \item{C}{Number of synthetic samples for each positive class samples.}
}
\description{
Resampling with SMOTE.
}
\details{
SMOTE (Chawla et al., 2002) is an oversampling method which creates links
between positive samples and nearest neighbors and generates synthetic
samples along that link.

It is well known that SMOTE is sensitive to noisy data. It may create more
noise.
}
\references{
Chawla, N. V., Bowyer, K. W., Hall, L. O., & Kegelmeyer, W. P. (2002). SMOTE:
synthetic minority over-sampling technique. Journal of artificial
intelligence research, 16, 321-357.
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
