% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_nb.R
\name{gaussian_nb}
\alias{gaussian_nb}
\title{Gaussian Naive Bayes Classifier}
\usage{
gaussian_nb(x, y, laplace = 0, sd_thresh = 1e-08, var_equal = FALSE)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{laplace}{asd}

\item{sd_thresh}{asd}

\item{var_equal}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Gaussian Naive Bayes Classifier.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
