% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_nb.R
\name{kernel_nb}
\alias{kernel_nb}
\title{Kernel Naive Bayes Classifier}
\usage{
kernel_nb(
  x,
  y,
  laplace = 0,
  bw_thresh = 1e-08,
  bw_method = "ucv",
  bw_adjust = 1
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{laplace}{asd}

\item{bw_thresh}{asd}

\item{bw_method}{asd}

\item{bw_adjust}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Kernel Naive Bayes Classifier.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
