% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locw_nb.R
\name{locw_nb}
\alias{locw_nb}
\title{Locally Weighted Naive Bayes}
\usage{
locw_nb(x, y, k = 50)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{k}{= asd,}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Locally Weighted Naive Bayes.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
