% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgaussian_b.R
\name{mvgaussian_b}
\alias{mvgaussian_b}
\title{Multivariate Gaussian Bayes Classifier}
\usage{
mvgaussian_b(x, y, laplace = 0, Sigma_equal = FALSE)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{laplace}{asd}

\item{Sigma_equal}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Multivariate Gaussian Bayes Classifier.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
