% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvkernel_b.R
\name{mvkernel_b}
\alias{mvkernel_b}
\title{Multivariate Kernel Bayes Classifier}
\usage{
mvkernel_b(
  x,
  y,
  laplace = 0,
  bw_thresh = 1e-08,
  bw_method = "nm",
  bw_adjust = 1
)
}
\arguments{
\item{x}{asd}

\item{y}{asd}

\item{laplace}{asd}

\item{bw_thresh}{asd}

\item{bw_method}{asd}

\item{bw_adjust}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Multivariate Kernel Bayes Classifier.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
