% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cl_copula_gaussian_b.R
\name{predict.cl_copula_gaussian_b}
\alias{predict.cl_copula_gaussian_b}
\title{Predict Clustered Copula Bayes Classifier with Gaussian marginals}
\usage{
\method{predict}{cl_copula_gaussian_b}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{asd}

\item{newdata}{asd}

\item{type}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Predicts class labels or probabilities for Clustered Copula Bayes Classifier with Gaussian marginals.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
