% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cl_gaussian_nb.R
\name{predict.cl_gaussian_nb}
\alias{predict.cl_gaussian_nb}
\title{Predict Clustered Gaussian Naive Bayes}
\usage{
\method{predict}{cl_gaussian_nb}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{asd}

\item{newdata}{asd}

\item{type}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Predicts class labels or probabilities for Clustered Gaussian Naive Bayes.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
