% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.copula_gaussian_b.R
\name{predict.copula_gaussian_b}
\alias{predict.copula_gaussian_b}
\title{Predict Copula Bayes Classifier with Gaussian marginals}
\usage{
\method{predict}{copula_gaussian_b}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{asd}

\item{newdata}{asd}

\item{type}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Predicts class labels or probabilities for Copula Bayes Classifier with Gaussian marginals.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
