% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cl_mvgaussian_b.R,
%   R/predict.gaussian_nb.R
\name{predict.cl_mvgaussian_b}
\alias{predict.cl_mvgaussian_b}
\alias{predict.gaussian_nb}
\title{Predict Clustered Multivariate Gaussian Bayes}
\usage{
\method{predict}{cl_mvgaussian_b}(object, newdata, type = "pred", ...)

\method{predict}{gaussian_nb}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{asd}

\item{newdata}{asd}

\item{type}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}

asd.
 \item{asd}{asd}
}
\description{
Predicts class labels or probabilities for Clustered Multivariate Gaussian Bayes.

Predicts class labels or probabilities for Gaussian Naive Bayes.
}
\details{
asd

asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
