% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.locw_nb.R
\name{predict.locw_nb}
\alias{predict.locw_nb}
\title{Predict Locally Weighted Naive Bayes}
\usage{
\method{predict}{locw_nb}(object, newdata, type = "prob", sd_thresh = 1e-08, laplace = 0)
}
\arguments{
\item{object}{asd}

\item{newdata}{asd}

\item{type}{asd}

\item{sd_thresh}{asd}

\item{laplace}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Predicts class labels or probabilities for Locally Weighted Naive Bayes.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
