% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mvgaussian_b.R
\name{predict.mvgaussian_b}
\alias{predict.mvgaussian_b}
\title{Predict Multivariate Gaussian Bayes}
\usage{
\method{predict}{mvgaussian_b}(object, newdata, type = "pred", ...)
}
\arguments{
\item{object}{asd}

\item{newdata}{asd}

\item{type}{asd}

\item{...}{asd}
}
\value{
asd.
 \item{asd}{asd}
}
\description{
Predicts class labels or probabilities for Multivariate Gaussian Bayes.
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
