% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowProd.R
\name{rowProd}
\alias{rowProd}
\title{Row products}
\usage{
rowProd(X, prior = rep(1, ncol(X)))
}
\arguments{
\item{X}{asd}

\item{prior}{asd}
}
\value{
asd.
 \item{asd}{asd.}
}
\description{
Row products
}
\details{
asd
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
